refactor: remove /verify-$PROVIDER routes
Now everything is under /verify route
