+++ apps/ui/src/routes/auction.ts
@@ -1,6 +1,5 @@
 import AuctionOverview from '@/views/Auction/Overview.vue';
 import AuctionVerify from '@/views/Auction/Verify.vue';
-import AuctionVerifyWithProvider from '@/views/Auction/VerifyWithProvider.vue';
 import Auction from '@/views/Auction.vue';
 import Auctions from '@/views/Auctions.vue';
 
@@ -16,19 +15,7 @@ export default [
     component: Auction,
     children: [
       { path: '', name: 'auction-overview', component: AuctionOverview },
-      { path: 'verify', name: 'auction-verify', component: AuctionVerify },
-      {
-        path: 'verify-zkpassport',
-        name: 'auction-verify-zkpassport',
-        component: AuctionVerifyWithProvider,
-        props: { provider: 'zkpassport' }
-      },
-      {
-        path: 'verify-sumsub',
-        name: 'auction-verify-sumsub',
-        component: AuctionVerifyWithProvider,
-        props: { provider: 'sumsub' }
-      }
+      { path: 'verify', name: 'auction-verify', component: AuctionVerify }
     ]
   }
 ];
